// Frontend Feature-add/HotelBookingFrontend_Fixed/src/app/models/room.model.ts
export class RoomModel {
  roomId: number = 0;
  hotelId: number = 0;
  roomNumber: number = 0;
  roomType: string = '';
  pricePerNight: number = 0;
  capacity: number = 0;
  isAvailable: boolean = true;
  imageUrl: string = '';
}

export class CreateRoomModel {
  hotelId: number = 0;
  roomNumber: number = 0;
  roomType: string = '';
  pricePerNight: number = 0;
  capacity: number = 0;
  imageUrl?: string;
}